Allow gendered badge titles
Normally, a badge can be attributed to members of both sexes.
However, in langauges where common nouns have a masculine and feminine form, we can tag users as

- "Redakteur oder Redakteurin" - This is confusing in every-day speech
- "Redakteurinnen und Redakteure" - This describes the group adequately, but is quite a mouthful, and sounds stilted
- "Redakteur" - This will make all progressives up in arms
- "Redakteurin" - This will make all conservatives up in arms
- "Redakteur/innen" - This will make all hardcorcore conservatives up in arms, and sounds stilted
- "RedakteurIn" - Also politically disputed.

Therefore, allow badges to have a specific name, depending on the gender, so we can simply label users (if we know the gender) as "Redakteur" or "Redakteurin".

See hhucn/adhocracy.hhu_theme#402 for our specific rationale